Update status code & error code
